@@ -1,7 +1,9 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
 import { Paper } from '@mui/material'
-import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
+import 'react-toastify/dist/ReactToastify.css'
+import 'react-toastify/dist/ReactToastify.min.css'
+import { ToastContainer } from 'react-toastify'
 
 import { Home } from './pages/Home'
 import NoMatch from './pages/NoMatch'
@@ -18,6 +20,9 @@ import { useSelector } from 'react-redux'
 const App = () => {
   let { darkMode: mode } = useSelector(selectDarkMode)
   const theme = mode ? darkTheme : lightTheme
+
+
+
 
   return (
     <ThemeProvider theme={theme}>
